rf: check import map if initial lookup fails
This commit adds an additional search in the package importMap if the
initial lookup fails.  If the importMap has a key matching the
dependency, then the package graph is searched for the corresponding
value of the importMap.

This was discovered when the `cmd/go/internal/modload` package
attempted to import `modfile` via `golang.org/x/mod/modfile` which did
not exist in the package graph.  However, the importMap contained the
entry:
  golang.org/x/mod/modfile" -> "cmd/vendor/golang.org/x/mod/modfile"

The vendored path was in the package graph, and therefore was used in
place of the non-vendored path.